+++ config/run/exec_test.go
@@ -9,35 +9,25 @@ import (
 	"github.com/rliebz/tusk/ui"
 )
 
-func TestExecCommand_output(t *testing.T) {
+func TestExecCommand(t *testing.T) {
+	command := "exit 0"
 
-	command := `
-echo one
->&2 echo two
->&2 echo three
-echo four
-`
-
-	bufExpected := new(bytes.Buffer)
-	ui.Stderr = log.New(bufExpected, "", 0)
+	stderrBuf := new(bytes.Buffer)
+	ui.Stderr = log.New(stderrBuf, "", 0)
 	ui.PrintCommand(command)
-	ui.PrintCommandOutput("one")
-	ui.PrintCommandOutput("two")
-	ui.PrintCommandOutput("three")
-	ui.PrintCommandOutput("four")
-	expected := bufExpected.String()
+	stderrExpected := stderrBuf.String()
 
-	bufActual := new(bytes.Buffer)
-	ui.Stderr = log.New(bufActual, "", 0)
+	stderrActualBuf := new(bytes.Buffer)
+	ui.Stderr = log.New(stderrActualBuf, "", 0)
 	if err := ExecCommand(command); err != nil {
 		t.Fatalf(`execCommand("%s"): unexpected err: %s`, command, err)
 	}
-	actual := bufActual.String()
+	stderrActual := stderrActualBuf.String()
 
-	if expected != actual {
-		t.Fatalf(
-			"execCommand(\"%s\"):\nexpected output:\n`%s`\nactual output:\n`%s`",
-			command, expected, actual,
+	if stderrExpected != stderrActual {
+		t.Errorf(
+			"execCommand(\"%s\"):\nexpected stderr:\n`%s`\nactual:\n`%s`",
+			command, stderrExpected, stderrActual,
 		)
 	}
 }